refactor: split ffi helper modules
