+++ crates/mcp-compressor-core/src/ffi/session.rs
@@ -1,0 +1,99 @@
+use crate::proxy::ToolProxyServer;
+use crate::server::{CompressedServer, CompressedServerConfig, ProxyTransformMode};
+use crate::Error;
+
+use super::dto::{
+    FfiBackendConfig, FfiCompressedSessionConfig, FfiCompressedSessionInfo, FfiTool,
+};
+
+pub struct FfiCompressedSession {
+    info: FfiCompressedSessionInfo,
+}
+
+impl FfiCompressedSession {
+    pub fn bridge_url(&self) -> &str {
+        &self.info.bridge_url
+    }
+
+    pub fn token(&self) -> &str {
+        &self.info.token
+    }
+
+    pub fn info(&self) -> FfiCompressedSessionInfo {
+        self.info.clone()
+    }
+}
+
+fn parse_ffi_transform_mode(value: Option<&str>) -> Result<ProxyTransformMode, Error> {
+    match value.unwrap_or("compressed-tools") {
+        "compressed-tools" | "compressed" | "normal" => Ok(ProxyTransformMode::CompressedTools),
+        "cli" | "cli-mode" => Ok(ProxyTransformMode::Cli),
+        "just-bash" | "just_bash" => Ok(ProxyTransformMode::JustBash),
+        other => Err(Error::Config(format!("invalid transform mode: {other}"))),
+    }
+}
+
+async fn compressed_session_from_server(
+    server: CompressedServer,
+) -> Result<FfiCompressedSession, Error> {
+    let frontend_tools = server
+        .list_frontend_tools()
+        .await?
+        .into_iter()
+        .map(FfiTool::from)
+        .collect();
+    let just_bash_providers = server
+        .just_bash_provider_specs()
+        .into_iter()
+        .map(Into::into)
+        .collect();
+    let proxy = ToolProxyServer::start(server).await?;
+    Ok(FfiCompressedSession {
+        info: FfiCompressedSessionInfo {
+            bridge_url: proxy.bridge_url().to_string(),
+            token: proxy.token_value().to_string(),
+            frontend_tools,
+            just_bash_providers,
+        },
+    })
+}
+
+pub async fn start_compressed_session(
+    config: FfiCompressedSessionConfig,
+    backends: Vec<FfiBackendConfig>,
+) -> Result<FfiCompressedSession, Error> {
+    let server = CompressedServer::connect_multi_stdio(
+        CompressedServerConfig {
+            level: config.compression_level.parse()?,
+            server_name: config.server_name,
+            include_tools: config.include_tools,
+            exclude_tools: config.exclude_tools,
+            toonify: config.toonify,
+            transform_mode: parse_ffi_transform_mode(config.transform_mode.as_deref())?,
+            ..CompressedServerConfig::default()
+        },
+        backends.into_iter().map(Into::into).collect(),
+    )
+    .await?;
+    compressed_session_from_server(server).await
+}
+
+pub async fn start_compressed_session_from_mcp_config(
+    config: FfiCompressedSessionConfig,
+    mcp_config_json: &str,
+) -> Result<FfiCompressedSession, Error> {
+    let server = CompressedServer::connect_mcp_config_json(
+        CompressedServerConfig {
+            level: config.compression_level.parse()?,
+            server_name: config.server_name,
+            include_tools: config.include_tools,
+            exclude_tools: config.exclude_tools,
+            toonify: config.toonify,
+            transform_mode: parse_ffi_transform_mode(config.transform_mode.as_deref())?,
+            ..CompressedServerConfig::default()
+        },
+        mcp_config_json,
+    )
+    .await?;
+    compressed_session_from_server(server).await
+}
