+++ crates/mcp-compressor-core/src/ffi/types.rs
@@ -1,349 +1,8 @@
-//! JSON-serializable DTOs and helpers for PyO3 / napi-rs language bindings.
-//!
-//! These are not a C ABI. They are intentionally plain Rust data-transfer
-//! objects that binding crates can expose idiomatically in Python and
-//! TypeScript while sharing the same core behavior.
-
-use std::path::PathBuf;
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-
-use crate::cli::parser::parse_argv;
-use crate::client_gen::cli::CliGenerator;
-use crate::client_gen::generator::{ClientGenerator, GeneratorConfig};
-use crate::client_gen::python::PythonGenerator;
-use crate::client_gen::typescript::TypeScriptGenerator;
-use crate::compression::engine::{CompressionEngine, Tool};
-use crate::compression::CompressionLevel;
-use crate::config::topology::MCPConfig;
-use crate::oauth::{
-    clear_oauth_store, list_oauth_stores, oauth_store_root, remember_oauth_store,
-    OAuthStoreIndexEntry,
-};
-use crate::proxy::ToolProxyServer;
-use crate::server::{
-    BackendServerConfig, CompressedServer, CompressedServerConfig, JustBashCommandSpec,
-    JustBashProviderSpec, ProxyTransformMode,
-};
-use crate::Error;
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FfiOAuthStoreEntry {
-    pub backend_name: String,
-    pub backend_uri: String,
-    pub store_dir: PathBuf,
-}
-
-pub fn oauth_store_path() -> PathBuf {
-    oauth_store_root()
-}
-
-pub fn remember_oauth_backend(
-    backend_uri: &str,
-    backend_name: &str,
-    store_dir: PathBuf,
-) -> Result<(), Error> {
-    remember_oauth_store(backend_uri, backend_name, &store_dir).map_err(Error::Io)
-}
-
-pub fn list_oauth_credentials() -> Result<Vec<FfiOAuthStoreEntry>, Error> {
-    list_oauth_stores()
-        .map(|entries| entries.into_iter().map(Into::into).collect())
-        .map_err(Error::Io)
-}
-
-pub fn clear_oauth_credentials(target: Option<&str>) -> Result<Vec<PathBuf>, Error> {
-    clear_oauth_store(target).map_err(Error::Io)
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FfiTool {
-    pub name: String,
-    pub description: Option<String>,
-    pub input_schema: Value,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FfiMcpServer {
-    pub name: String,
-    pub command: String,
-    pub args: Vec<String>,
-    pub env: Vec<(String, String)>,
-    pub cli_prefix: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FfiJustBashProviderSpec {
-    pub provider_name: String,
-    pub help_tool_name: String,
-    pub tools: Vec<FfiJustBashCommandSpec>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FfiJustBashCommandSpec {
-    pub command_name: String,
-    pub backend_tool_name: String,
-    pub description: Option<String>,
-    pub input_schema: Value,
-    pub invoke_tool_name: String,
-}
-
-pub fn compress_tool_listing(level: CompressionLevel, tools: Vec<FfiTool>) -> String {
-    let tools = tools.into_iter().map(Into::into).collect::<Vec<Tool>>();
-    CompressionEngine::new(level).format_listing(&tools)
-}
-
-pub fn format_tool_schema_response(tool: FfiTool) -> String {
-    CompressionEngine::format_schema_response(&tool.into())
-}
-
-pub fn parse_tool_argv(tool: FfiTool, argv: Vec<String>) -> Result<Value, Error> {
-    parse_argv(&argv, &tool.into())
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FfiGeneratorConfig {
-    pub cli_name: String,
-    pub bridge_url: String,
-    pub token: String,
-    pub tools: Vec<FfiTool>,
-    pub session_pid: u32,
-    pub output_dir: PathBuf,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub enum FfiClientArtifactKind {
-    Cli,
-    Python,
-    TypeScript,
-}
-
-pub fn generate_client_artifacts(
-    kind: FfiClientArtifactKind,
-    config: FfiGeneratorConfig,
-) -> Result<Vec<PathBuf>, Error> {
-    let config = GeneratorConfig::from(config);
-    match kind {
-        FfiClientArtifactKind::Cli => CliGenerator.generate(&config),
-        FfiClientArtifactKind::Python => PythonGenerator.generate(&config),
-        FfiClientArtifactKind::TypeScript => TypeScriptGenerator.generate(&config),
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FfiBackendConfig {
-    pub name: String,
-    pub command_or_url: String,
-    #[serde(default)]
-    pub args: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FfiCompressedSessionConfig {
-    pub compression_level: String,
-    pub server_name: Option<String>,
-    #[serde(default)]
-    pub include_tools: Vec<String>,
-    #[serde(default)]
-    pub exclude_tools: Vec<String>,
-    #[serde(default)]
-    pub toonify: bool,
-    pub transform_mode: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FfiCompressedSessionInfo {
-    pub bridge_url: String,
-    pub token: String,
-    pub frontend_tools: Vec<FfiTool>,
-    pub just_bash_providers: Vec<FfiJustBashProviderSpec>,
-}
-
-pub struct FfiCompressedSession {
-    info: FfiCompressedSessionInfo,
-}
-
-impl FfiCompressedSession {
-    pub fn bridge_url(&self) -> &str {
-        &self.info.bridge_url
-    }
-
-    pub fn token(&self) -> &str {
-        &self.info.token
-    }
-
-    pub fn info(&self) -> FfiCompressedSessionInfo {
-        self.info.clone()
-    }
-}
-
-fn parse_ffi_transform_mode(value: Option<&str>) -> Result<ProxyTransformMode, Error> {
-    match value.unwrap_or("compressed-tools") {
-        "compressed-tools" | "compressed" | "normal" => Ok(ProxyTransformMode::CompressedTools),
-        "cli" | "cli-mode" => Ok(ProxyTransformMode::Cli),
-        "just-bash" | "just_bash" => Ok(ProxyTransformMode::JustBash),
-        other => Err(Error::Config(format!("invalid transform mode: {other}"))),
-    }
-}
-
-async fn compressed_session_from_server(
-    server: CompressedServer,
-) -> Result<FfiCompressedSession, Error> {
-    let frontend_tools = server
-        .list_frontend_tools()
-        .await?
-        .into_iter()
-        .map(Into::into)
-        .collect();
-    let just_bash_providers = server
-        .just_bash_provider_specs()
-        .into_iter()
-        .map(Into::into)
-        .collect();
-    let proxy = ToolProxyServer::start(server).await?;
-    Ok(FfiCompressedSession {
-        info: FfiCompressedSessionInfo {
-            bridge_url: proxy.bridge_url().to_string(),
-            token: proxy.token_value().to_string(),
-            frontend_tools,
-            just_bash_providers,
-        },
-    })
-}
-
-pub async fn start_compressed_session(
-    config: FfiCompressedSessionConfig,
-    backends: Vec<FfiBackendConfig>,
-) -> Result<FfiCompressedSession, Error> {
-    let server = CompressedServer::connect_multi_stdio(
-        CompressedServerConfig {
-            level: config.compression_level.parse()?,
-            server_name: config.server_name,
-            include_tools: config.include_tools,
-            exclude_tools: config.exclude_tools,
-            toonify: config.toonify,
-            transform_mode: parse_ffi_transform_mode(config.transform_mode.as_deref())?,
-            ..CompressedServerConfig::default()
-        },
-        backends.into_iter().map(Into::into).collect(),
-    )
-    .await?;
-    compressed_session_from_server(server).await
-}
-
-pub async fn start_compressed_session_from_mcp_config(
-    config: FfiCompressedSessionConfig,
-    mcp_config_json: &str,
-) -> Result<FfiCompressedSession, Error> {
-    let server = CompressedServer::connect_mcp_config_json(
-        CompressedServerConfig {
-            level: config.compression_level.parse()?,
-            server_name: config.server_name,
-            include_tools: config.include_tools,
-            exclude_tools: config.exclude_tools,
-            toonify: config.toonify,
-            transform_mode: parse_ffi_transform_mode(config.transform_mode.as_deref())?,
-            ..CompressedServerConfig::default()
-        },
-        mcp_config_json,
-    )
-    .await?;
-    compressed_session_from_server(server).await
-}
-
-pub fn parse_mcp_config(config_json: &str) -> Result<Vec<FfiMcpServer>, Error> {
-    let config = MCPConfig::from_json(config_json)?;
-    Ok(config
-        .server_names()
-        .into_iter()
-        .filter_map(|name| {
-            let server = config.server(&name)?;
-            Some(FfiMcpServer {
-                cli_prefix: config.cli_prefix(&name),
-                name,
-                command: server.command.clone(),
-                args: server.args.clone(),
-                env: server
-                    .env
-                    .iter()
-                    .map(|(key, value)| (key.clone(), value.clone()))
-                    .collect(),
-            })
-        })
-        .collect())
-}
-
-impl From<OAuthStoreIndexEntry> for FfiOAuthStoreEntry {
-    fn from(value: OAuthStoreIndexEntry) -> Self {
-        Self {
-            backend_name: value.name,
-            backend_uri: value.uri,
-            store_dir: value.store_dir.into(),
-        }
-    }
-}
-
-impl From<FfiBackendConfig> for BackendServerConfig {
-    fn from(value: FfiBackendConfig) -> Self {
-        BackendServerConfig::new(value.name, value.command_or_url, value.args)
-    }
-}
-
-impl From<FfiGeneratorConfig> for GeneratorConfig {
-    fn from(value: FfiGeneratorConfig) -> Self {
-        Self {
-            cli_name: value.cli_name,
-            bridge_url: value.bridge_url,
-            token: value.token,
-            tools: value.tools.into_iter().map(Into::into).collect(),
-            session_pid: value.session_pid,
-            output_dir: value.output_dir,
-        }
-    }
-}
-
-impl From<FfiTool> for Tool {
-    fn from(value: FfiTool) -> Self {
-        Tool::new(value.name, value.description, value.input_schema)
-    }
-}
-
-impl From<Tool> for FfiTool {
-    fn from(value: Tool) -> Self {
-        Self {
-            name: value.name,
-            description: value.description,
-            input_schema: value.input_schema,
-        }
-    }
-}
-
-impl From<JustBashProviderSpec> for FfiJustBashProviderSpec {
-    fn from(value: JustBashProviderSpec) -> Self {
-        Self {
-            provider_name: value.provider_name,
-            help_tool_name: value.help_tool_name,
-            tools: value.tools.into_iter().map(Into::into).collect(),
-        }
-    }
-}
-
-impl From<JustBashCommandSpec> for FfiJustBashCommandSpec {
-    fn from(value: JustBashCommandSpec) -> Self {
-        Self {
-            command_name: value.command_name,
-            backend_tool_name: value.backend_tool_name,
-            description: value.description,
-            input_schema: value.input_schema,
-            invoke_tool_name: value.invoke_tool_name,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use serde_json::json;
+    use super::super::*;
+    use crate::compression::CompressionLevel;
+    use serde_json::{json, Value};
 
     fn sample_tool() -> FfiTool {
         FfiTool {
